refactor: return number of watchtowers as u32
Watchtower indices are already u32,
so making the number of watchtowers u32 feels natural.
Coverting from u32 to usize feels safer than converting
from usize to u32.

+++ crates/tx-graph2/src/transactions/contest.rs
@@ -66,7 +66,7 @@ impl ContestTx {
         let value_out = proof_connector.value()
             + payout_connector.value()
             + slash_connector.value()
-            + counterproof_output.value() * claim_contest_connector.n_watchtowers() as u64;
+            + counterproof_output.value() * u64::from(claim_contest_connector.n_watchtowers());
         debug_assert!(
             claim_contest_connector.value() == value_out,
             "tx should have zero fees (value in = {}, value out = {value_out}",
@@ -97,7 +97,7 @@ impl ContestTx {
         ];
         output.extend(std::iter::repeat_n(
             counterproof_output.tx_out(),
-            claim_contest_connector.n_watchtowers(),
+            claim_contest_connector.n_watchtowers() as usize,
         ));
         output.push(cpfp_connector.tx_out());
